TIKA-4717 - promote external parser v2
