+++ docs/modules/ROOT/examples/external-parser-ffmpeg.json
@@ -1,0 +1,1 @@
+../../../../tika-serialization/src/test/resources/configs/external-parser-ffmpeg.json
