+++ docs/modules/ROOT/examples/external-parser-multi.json
@@ -1,0 +1,1 @@
+../../../../tika-serialization/src/test/resources/configs/external-parser-multi.json
