+++ docs/modules/ROOT/pages/configuration/parsers/external-parser.adoc
@@ -1,0 +1,154 @@
+//
+// Licensed to the Apache Software Foundation (ASF) under one or more
+// contributor license agreements.  See the NOTICE file distributed with
+// this work for additional information regarding copyright ownership.
+// The ASF licenses this file to You under the Apache License, Version 2.0
+// (the "License"); you may not use this file except in compliance with
+// the License.  You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+= External Parser Configuration
+
+The `ExternalParser` allows Tika to delegate parsing to external command-line
+programs such as `ffmpeg`, `exiftool`, or `sox`. Each external parser is
+configured via JSON and must be explicitly enabled -- Tika 4.x does not
+auto-discover external tools at startup.
+
+== Key Concepts
+
+=== Lazy Check
+
+Each external parser can declare a `checkCommandLine` that verifies the tool
+is installed. The check runs lazily on first use (not at startup), and if the
+tool is not found, the parser silently disables itself.
+
+=== Output Parser vs Stderr Parser
+
+External tools write useful output to different streams:
+
+* **`outputParser`** -- processes stdout (or the output file). Use this for tools
+  like `exiftool` that write structured output to stdout.
+* **`stderrParser`** -- processes stderr. Use this for tools like `ffmpeg` and
+  `sox` that write metadata to stderr.
+
+Both accept any Tika parser; `regex-capture-parser` is the most common choice
+for extracting metadata via regex patterns.
+
+== Configuration Options
+
+[cols="1,1,3"]
+|===
+|Field |Type |Description
+
+|`commandLine`
+|`List<String>`
+|The command and arguments to run. Use `${INPUT_FILE}` and `${OUTPUT_FILE}` tokens for file paths.
+
+|`checkCommandLine`
+|`List<String>`
+|Optional. Command to verify the tool is installed (e.g., `["ffmpeg", "-version"]`).
+
+|`checkErrorCodes`
+|`List<Integer>`
+|Exit codes that indicate the tool is not available. Default: `[127]`.
+
+|`outputParser`
+|Parser config
+|Optional. Parser to process stdout or the output file.
+
+|`stderrParser`
+|Parser config
+|Optional. Parser to process stderr (for metadata extraction).
+
+|`returnStdout`
+|`boolean`
+|Store raw stdout in metadata. Default: `false`.
+
+|`returnStderr`
+|`boolean`
+|Store raw stderr in metadata. Default: `true`.
+
+|`timeoutMs`
+|`long`
+|Process timeout in milliseconds. Default: `60000`.
+
+|`maxStdOut`
+|`int`
+|Maximum stdout bytes to capture. Default: `10000`.
+
+|`maxStdErr`
+|`int`
+|Maximum stderr bytes to capture. Default: `10000`.
+|===
+
+== Examples
+
+=== Exiftool (metadata from stdout)
+
+Extracts metadata from media files using `exiftool`. The `outputParser` uses
+`regex-capture-parser` to extract key-value pairs from exiftool's stdout.
+
+[source,json]
+----
+include::example$external-parser-exiftool.json[]
+----
+icon:github[] https://github.com/apache/tika/blob/main/tika-serialization/src/test/resources/configs/external-parser-exiftool.json[View source on GitHub]
+
+=== FFmpeg (metadata from stderr)
+
+Extracts audio/video metadata from `ffmpeg -i` output. FFmpeg writes metadata
+to stderr, so this uses `stderrParser` instead of `outputParser`.
+
+[source,json]
+----
+include::example$external-parser-ffmpeg.json[]
+----
+icon:github[] https://github.com/apache/tika/blob/main/tika-serialization/src/test/resources/configs/external-parser-ffmpeg.json[View source on GitHub]
+
+=== Sox (audio metadata from stderr)
+
+Extracts audio metadata using `sox --info`. Like FFmpeg, Sox writes to stderr.
+
+[source,json]
+----
+include::example$external-parser-sox.json[]
+----
+icon:github[] https://github.com/apache/tika/blob/main/tika-serialization/src/test/resources/configs/external-parser-sox.json[View source on GitHub]
+
+=== Multiple External Parsers
+
+You can configure multiple external parsers in a single config file. Each
+handles different MIME types via `_mime-include`. Here FFmpeg handles video
+files while exiftool handles PDFs:
+
+[source,json]
+----
+include::example$external-parser-multi.json[]
+----
+icon:github[] https://github.com/apache/tika/blob/main/tika-serialization/src/test/resources/configs/external-parser-multi.json[View source on GitHub]
+
+== Changes from 3.x
+
+In Tika 3.x, external parsers were configured via XML (`tika-external-parsers.xml`)
+and auto-discovered at startup. The `CompositeExternalParser` would fork
+a process for each configured tool on every Tika initialization to check
+if the tool was available.
+
+In Tika 4.x:
+
+* External parsers must be explicitly configured in JSON -- no auto-discovery.
+* The `checkCommandLine` runs lazily on first use, not at startup.
+* The `stderrParser` field replaces the inline regex-on-stderr metadata extraction.
+* The `external2` package has been renamed back to `external`.
+* `CompositeExternalParser`, `ExternalParsersFactory`, and the XML config
+  reader have been removed.
+
+See xref:migration-to-4x/migrating-to-4x.adoc[Migrating to 4.x] for general migration guidance.
