--- tika-core/src/main/resources/org/apache/tika/parser/external/tika-external-parsers.xml
@@ -1,117 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one or more
-  contributor license agreements.  See the NOTICE file distributed with
-  this work for additional information regarding copyright ownership.
-  The ASF licenses this file to You under the Apache License, Version 2.0
-  (the "License"); you may not use this file except in compliance with
-  the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<!--
-  Description: This xml file defines external commands to be run by Tika
-  as parsers.
--->
-<external-parsers>
-  <!-- This example uses ffmpeg for video metadata extraction -->
-  <parser>
-     <check>
-       <command>ffmpeg -version</command>
-       <error-codes>126,127</error-codes>
-     </check>
-     <command>ffmpeg -i ${INPUT}</command>
-     <mime-types>
-       <mime-type>video/avi</mime-type>
-       <mime-type>video/mpeg</mime-type>
-       <mime-type>video/x-msvideo</mime-type>
-     </mime-types>
-     <metadata>
-       <match key="xmpDM:audioSampleRate">\s*Stream.*:.+Audio:.*,\s+(\d+)\s+Hz,.*</match>
-       <match key="xmpDM:audioChannelType">\s*Stream.*:.+Audio:.*\d+\s+Hz,\s+(\d{1,2})\s+channels.*</match>
-       <match key="xmpDM:audioCompressor">\s*Stream.*:.+Audio:\s+([A-Za-z0-9_\(\)/\[\] ]+),.*</match>
-       <match key="xmpDM:duration">\s*Duration:\s*([0-9:\.]+),.*</match>
-       <match key="xmpDM:fileDataRate">\s*Duration:.*,\s*bitrate:\s+([0-9A-Za-z/ ]+).*</match>
-       <match key="xmpDM:videoColorSpace">\s*Stream.*:\s+Video:\s+[A-Za-z0-9\(\)/ ]+,\s+([A-Za-z0-9\(\) ,]+),\s+[0-9x]+,.*</match>
-       <match key="xmpDM:videoCompressor">\s*Stream.*:\s+Video:\s+([A-Za-z0-9\(\)/ ]+),.*</match>
-       <match key="xmpDM:videoFrameRate">\s*Stream.*:\s+Video:.*,\s+([0-9]+)\s+fps,.*</match>
-       <match key="encoder">\s*encoder\s*\:\s*(\w+).*</match>
-       <match key="videoResolution">\s*Stream.*:\s+Video:.*,\s+([0-9x]+),.*</match>
-     </metadata>
-  </parser>
-  <parser>
-     <check>
-       <command>exiftool -ver</command>
-       <error-codes>126,127</error-codes>
-     </check>
-     <command>env FOO=${OUTPUT} exiftool ${INPUT}</command>
-     <mime-types>
-       <mime-type>video/avi</mime-type>
-       <mime-type>video/mpeg</mime-type>
-       <mime-type>video/x-msvideo</mime-type>
-       <mime-type>video/mp4</mime-type>
-     </mime-types>
-     <metadata>
-       <match>\s*([A-Za-z0-9/ \(\)]+\S{1})\s+:\s+([A-Za-z0-9\(\)\[\] \:\-\.]+)\s*</match>
-     </metadata>
-  </parser>
-  <parser>
-     <check>
-       <command>sox --version</command>
-       <error-codes>126,127</error-codes>
-     </check>
-     <command>env FOO=${OUTPUT} sox --info ${INPUT}</command>
-     <mime-types>
-       <mime-type>audio/3gpp</mime-type>
-       <mime-type>audio/3gpp2</mime-type>
-       <mime-type>audio/aac</mime-type>
-       <mime-type>audio/ac3</mime-type>
-       <mime-type>audio/basic</mime-type>
-       <mime-type>audio/L24</mime-type>
-       <mime-type>audio/mid</mime-type>
-       <mime-type>audio/mpeg</mime-type>
-       <mime-type>audio/mpeg3</mime-type>
-       <mime-type>audio/x-mpeg-3</mime-type>
-       <mime-type>audio/mpeg4-generic</mime-type>
-       <mime-type>audio/mp4</mime-type>
-       <mime-type>audio/mp3</mime-type>
-       <mime-type>audio/x-aiff</mime-type>
-       <mime-type>audio/PCMA</mime-type>
-       <mime-type>audio/PCMA-WB</mime-type>
-       <mime-type>audio/PCMU</mime-type>
-       <mime-type>audio/PCMU-WB</mime-type>
-       <mime-type>audio/ogg</mime-type>
-       <mime-type>audio/vorbis</mime-type>
-       <mime-type>audio/vnd.wav</mime-type>
-       <mime-type>audio/vnd.wave</mime-type>
-       <mime-type>audio/vnd.rn-realaudio</mime-type>
-       <mime-type>audio/wav</mime-type>
-       <mime-type>audio/wave</mime-type>
-       <mime-type>audio/x-wav</mime-type>
-     </mime-types>
-     <metadata>
-       <!-- Channels       : 1 -->
-       <match key="xmpDM:audioChannelType">\s*Channels.*:\s+(\d+)\s*</match>
-       <!-- Sample Rate    : 44100 -->
-       <match key="xmpDM:audioSampleRate">\s*Sample Rate.*:\s+(\d+)\s*</match>
-       <!-- Precision      : 16-bit -->
-       <match key="xmpDM:audioSampleType">\s*Precision.*:\s+([\d\w-]+)\s*</match>
-       <!-- Duration       : 00:00:02.50 = 110298 samples = 187.582 CDDA sectors -->
-       <match key="xmpDM:duration">\s*Duration.*:\s+([\d:\.]+)\s*</match>
-       <!-- File Size      : 221k -->
-       <match key="File Size">\s*File Size.*:\s+([\d\w]+)\s*</match>
-       <!-- Bit Rate       : 706k -->
-       <match key="xmpDM:fileDataRate">\s*Bit Rate.*:\s+([\d\w]+)\s*</match>
-       <!-- Sample Encoding: 16-bit Signed Integer PCM -->
-       <match key="Sample Encoding">\s*Sample Encoding.*:\s+(.*)\s*</match>
-       <!--  Comment        : 'Comment=Processed by SoX' -->
-       <match key="xmpDM:logComment">\s*Comment.*:\s+(.*)\s*</match>
-     </metadata>
-  </parser>
-</external-parsers>
