+++ tika-detectors/tika-detector-siegfried/src/main/java/org/apache/tika/detect/siegfried/SiegfriedDetector.java
@@ -35,7 +35,6 @@ import org.apache.tika.metadata.Metadata;
 import org.apache.tika.metadata.Property;
 import org.apache.tika.mime.MediaType;
 import org.apache.tika.parser.ParseContext;
-import org.apache.tika.parser.external.ExternalParser;
 import org.apache.tika.utils.FileProcessResult;
 import org.apache.tika.utils.ProcessUtils;
 import org.apache.tika.utils.StringUtils;
@@ -170,7 +169,7 @@ public class SiegfriedDetector implements Detector {
 
     public static boolean checkHasSiegfried(String siegfriedCommandPath) {
         String[] commandline = new String[]{siegfriedCommandPath, "-version"};
-        return ExternalParser.check(commandline);
+        return ProcessUtils.checkCommand(commandline);
     }
 
     /**
