+++ tika-parsers/tika-parsers-extended/tika-parser-scientific-package/src/test/java/org/apache/tika/parser/scientific/integration/TestParsers.java
@@ -30,10 +30,8 @@ import org.junit.jupiter.api.Test;
 
 import org.apache.tika.exception.TikaException;
 import org.apache.tika.mime.MediaType;
-import org.apache.tika.parser.CompositeParser;
 import org.apache.tika.parser.DefaultParser;
 import org.apache.tika.parser.Parser;
-import org.apache.tika.parser.external.CompositeExternalParser;
 import org.apache.tika.parser.ocr.TesseractOCRParser;
 
 /**
@@ -58,11 +56,6 @@ public class TestParsers {
         }
 
         int checked = 0;
-        //The initial lists were developed with exiftool installed.  We have since
-        //modified the 2.4.1-* files to act as if no exiftool is installed.
-        //However, on systems with ffmpeg or exiftool installed, we need
-        //to override those file formats
-        CompositeParser externalParser = (CompositeParser) new CompositeExternalParser();
         try (BufferedReader reader =
                      new BufferedReader(new InputStreamReader(
                              getClass().getResourceAsStream(path241),
@@ -73,10 +66,6 @@ public class TestParsers {
                 String mediaType = data[0];
                 String parserClass = data[1];
 
-                Parser external = externalParser.getParsers().get(MediaType.parse(mediaType));
-                if (external != null) {
-                    parserClass = externalParser.getClass().toString();
-                }
                 assertEquals(parserClass, currentDefault.get(mediaType),
                         "for mediaType '" + mediaType + "'");
                 checked++;
