+++ tika-parsers/tika-parsers-standard/tika-parsers-standard-package/src/test/java/org/apache/tika/parser/AutoDetectParserTest.java
@@ -43,7 +43,6 @@ import org.apache.tika.metadata.Metadata;
 import org.apache.tika.metadata.TikaCoreProperties;
 import org.apache.tika.metadata.XMPDM;
 import org.apache.tika.mime.MediaType;
-import org.apache.tika.parser.external.CompositeExternalParser;
 import org.apache.tika.parser.ogg.FlacParser;
 import org.apache.tika.parser.ogg.OpusParser;
 import org.apache.tika.parser.ogg.VorbisParser;
@@ -407,11 +406,6 @@ public class AutoDetectParserTest extends TikaTest {
         }
     }
 
-    @Test
-    public void testExternalParserIsLoaded() {
-        Parser p = find((CompositeParser) AUTO_DETECT_PARSER, CompositeExternalParser.class);
-        assertNotNull(p);
-    }
 
     @Test
     public void testWriteLimit() throws Exception {
