+++ tika-serialization/src/test/resources/configs/external-parser-ffmpeg.json
@@ -1,0 +1,34 @@
+{
+  "parsers": [
+    {
+      "external-parser": {
+        "_mime-include": [
+          "video/avi",
+          "video/mpeg",
+          "video/x-msvideo"
+        ],
+        "commandLine": ["ffmpeg", "-i", "${INPUT_FILE}"],
+        "checkCommandLine": ["ffmpeg", "-version"],
+        "checkErrorCodes": [126, 127],
+        "returnStderr": true,
+        "maxStdErr": 20000,
+        "stderrParser": {
+          "regex-capture-parser": {
+            "captureMap": {
+              "xmpDM:audioSampleRate": "\\s*Stream.*:.+Audio:.*,\\s+(\\d+)\\s+Hz,.*",
+              "xmpDM:audioChannelType": "\\s*Stream.*:.+Audio:.*\\d+\\s+Hz,\\s+(\\d{1,2})\\s+channels.*",
+              "xmpDM:audioCompressor": "\\s*Stream.*:.+Audio:\\s+([A-Za-z0-9_\\(\\)/\\[\\] ]+),.*",
+              "xmpDM:duration": "\\s*Duration:\\s*([0-9:\\.]+),.*",
+              "xmpDM:fileDataRate": "\\s*Duration:.*,\\s*bitrate:\\s+([0-9A-Za-z/ ]+).*",
+              "xmpDM:videoColorSpace": "\\s*Stream.*:\\s+Video:\\s+[A-Za-z0-9\\(\\)/ ]+,\\s+([A-Za-z0-9\\(\\) ,]+),\\s+[0-9x]+,.*",
+              "xmpDM:videoCompressor": "\\s*Stream.*:\\s+Video:\\s+([A-Za-z0-9\\(\\)/ ]+),.*",
+              "xmpDM:videoFrameRate": "\\s*Stream.*:\\s+Video:.*,\\s+([0-9]+)\\s+fps,.*",
+              "encoder": "\\s*encoder\\s*\\:\\s*(\\w+).*",
+              "videoResolution": "\\s*Stream.*:\\s+Video:.*,\\s+([0-9x]+),.*"
+            }
+          }
+        }
+      }
+    }
+  ]
+}
